+++ cmd/genji/go.mod
@@ -4,13 +4,10 @@ go 1.15
 
 require (
 	github.com/agnivade/levenshtein v1.1.0
-	github.com/c-bata/go-prompt v0.2.3
-	github.com/dgraph-io/badger/v2 v2.2007.1
+	github.com/c-bata/go-prompt v0.2.5
+	github.com/dgraph-io/badger/v2 v2.2007.2
 	github.com/genjidb/genji v0.8.0
 	github.com/genjidb/genji/engine/badgerengine v0.8.0
-	github.com/mattn/go-runewidth v0.0.9 // indirect
-	github.com/mattn/go-tty v0.0.3 // indirect
-	github.com/pkg/term v0.0.0-20200520122047-c3ffed290a03 // indirect
 	github.com/stretchr/testify v1.6.1
 	github.com/urfave/cli/v2 v2.2.0
 )
