+++ document/iterator.go
@@ -1,7 +1,9 @@
 package document
 
 import (
+	"bufio"
 	"errors"
+	"io"
 )
 
 // ErrStreamClosed is used to indicate that a stream must be closed.
@@ -41,6 +43,52 @@ type IteratorFunc func(func(d Document) error) error
 // Iterate calls f(fn).
 func (f IteratorFunc) Iterate(fn func(d Document) error) error {
 	return f(fn)
+}
+
+// IteratorToJSON encodes all the documents of an iterator to JSON stream.
+func IteratorToJSON(w io.Writer, s Iterator) error {
+	buf := bufio.NewWriter(w)
+	defer buf.Flush()
+
+	return s.Iterate(func(d Document) error {
+		data, err := jsonDocument{d}.MarshalJSON()
+		if err != nil {
+			return err
+		}
+
+		_, err = buf.Write(data)
+		return err
+	})
+}
+
+// IteratorToJSONArray encodes all the documents of an iterator to a JSON array.
+func IteratorToJSONArray(w io.Writer, s Iterator) error {
+	buf := bufio.NewWriter(w)
+
+	buf.WriteByte('[')
+
+	first := true
+	err := s.Iterate(func(d Document) error {
+		if !first {
+			buf.WriteString(", ")
+		} else {
+			first = false
+		}
+
+		data, err := jsonDocument{d}.MarshalJSON()
+		if err != nil {
+			return err
+		}
+
+		_, err = buf.Write(data)
+		return err
+	})
+	if err != nil {
+		return err
+	}
+
+	buf.WriteByte(']')
+	return buf.Flush()
 }
 
 // Stream reads documents of an iterator one by one and passes them
