+++ document/iterator_test.go
@@ -1,11 +1,15 @@
 package document_test
 
 import (
+	"bytes"
+	"encoding/json"
 	"fmt"
 	"log"
+	"testing"
 
 	"github.com/genjidb/genji"
 	"github.com/genjidb/genji/document"
+	"github.com/stretchr/testify/require"
 )
 
 func ExampleStream_First() {
@@ -144,4 +148,20 @@ func ExampleStream_Iterate() {
 	// 9 foo9 90 map[city:Lyon zipcode:69009]
 	// {10 foo10 100 {Lyon 69010}}
 	// 10 foo10 100 map[city:Lyon zipcode:69010]
+}
+
+func TestIteratorToJSONArray(t *testing.T) {
+	var docs []document.Document
+	for i := 0; i < 3; i++ {
+		fb := document.NewFieldBuffer()
+		err := json.Unmarshal([]byte(fmt.Sprintf(`{"a": %d}`, i)), fb)
+		require.NoError(t, err)
+		docs = append(docs, fb)
+	}
+
+	it := document.NewIterator(docs...)
+	var buf bytes.Buffer
+	err := document.IteratorToJSONArray(&buf, it)
+	require.NoError(t, err)
+	require.Equal(t, `[{"a": 0}, {"a": 1}, {"a": 2}]`, buf.String())
 }
