+++ go.mod
@@ -3,6 +3,7 @@ module github.com/genjidb/genji
 go 1.15
 
 require (
+	github.com/buger/jsonparser v1.0.0
 	github.com/google/btree v1.0.0
 	github.com/stretchr/testify v1.6.1
 	github.com/vmihailenco/msgpack/v5 v5.0.0-beta.1
